+++ lib/src/outbound/outbound.freezed.dart
@@ -12,7 +12,7 @@ part of 'outbound.dart';
 T _$identity<T>(T value) => value;
 
 final _privateConstructorUsedError = UnsupportedError(
-    'It seems like you constructed your class using `MyClass._()`. This constructor is only meant to be used by freezed and you are not supposed to need it nor use it.\nPlease check the documentation here for more information: https://github.com/rrousselGit/freezed#adding-getters-and-methods-to-our-models');
+    'It seems like you constructed your class using `MyClass._()`. This constructor is only meant to be used by freezed and you are not supposed to need it nor use it.\nPlease check the documentation here for more information: https://github.com/rrousselGit/freezed#custom-getters-and-methods');
 
 OutboundProtocol _$OutboundProtocolFromJson(Map<String, dynamic> json) {
   return _OutboundProtocol.fromJson(json);
@@ -276,7 +276,7 @@ class _$OutboundProtocolImpl implements _OutboundProtocol {
   }
 
   @override
-  bool operator ==(Object other) {
+  bool operator ==(dynamic other) {
     return identical(this, other) ||
         (other.runtimeType == runtimeType &&
             other is _$OutboundProtocolImpl &&
@@ -799,7 +799,7 @@ class _$OutboundServiceImpl implements _OutboundService {
   }
 
   @override
-  bool operator ==(Object other) {
+  bool operator ==(dynamic other) {
     return identical(this, other) ||
         (other.runtimeType == runtimeType &&
             other is _$OutboundServiceImpl &&
@@ -1059,7 +1059,7 @@ class _$OutboundStructureImpl implements _OutboundStructure {
   }
 
   @override
-  bool operator ==(Object other) {
+  bool operator ==(dynamic other) {
     return identical(this, other) ||
         (other.runtimeType == runtimeType &&
             other is _$OutboundStructureImpl &&